implement sparse tree construction using unordered_map and fix FMM solvers
- Refactor tree hierarchy in Tree class to use std::unordered_map for sparse depth levels
- Implement Node::ChildIterator inside Node class to iterate over unordered_map child keys
- Add self-leaf particle P2P direct interaction loop and local_set null guards in Leaf::compute_acceleration
- Fix initialization order in Particle constructor (position before acceleration)

+++ src/system_updaters/tree_components/morton_code.cpp
@@ -6,8 +6,8 @@
 #include "morton_code.hpp"
 #include <cstring>
 
-size_t Morton::encode(const unsigned int *cartesian_coordinates, unsigned int dim){
-  unsigned int code = 0;
+uint64_t Morton::encode(const unsigned int *cartesian_coordinates, unsigned int dim){
+  uint64_t code = 0;
   unsigned int temp_cartesian_coordinates[dim];
   std::memcpy(temp_cartesian_coordinates, cartesian_coordinates, sizeof(unsigned int) * dim);
   for (unsigned int i = 0; i < dim; ++i){
@@ -21,7 +21,7 @@ size_t Morton::encode(const unsigned int *cartesian_coordinates, unsigned int di
   return code;
 }
 
-void Morton::decode(unsigned int *cartesian_coordinates, unsigned int dim, size_t code){
+void Morton::decode(unsigned int *cartesian_coordinates, unsigned int dim, uint64_t code){
   for (unsigned int i = 0; i < dim; ++i)
     cartesian_coordinates[i] = 0;
